Updated types for gil.

+++ service/JavaService/src/main/java/io/eaglejs/ffxi/mapper/PlayerMapper.java
@@ -121,7 +121,7 @@ public class PlayerMapper {
         player.setNationRank(document.getInteger("nationRank"));
         player.setStatus(document.getInteger("status"));
         player.setLastOnline(document.getLong("lastOnline"));
-        player.setGil(document.getLong("gil"));
+        player.setGil(document.getInteger("gil"));
         player.setHpp(document.getInteger("hpp"));
         player.setMpp(document.getInteger("mpp"));
         player.setTp(document.getInteger("tp"));
